refactor(infrastructure): PR-7 commit 3 — db.py 188 + migrations/ 搬 infrastructure/
  studio/infrastructure/db.py         188 行整搬迁（git rename 100%）
  studio/infrastructure/migrations/   9 子文件整体搬迁（git rename 100%）
                                      含 __init__.py + _v2..._v9
  studio/db.py                        11 行 shim（sys.modules 别名）

connection/tasks/settings 3-way 拆同 secrets：planning 原方案延后到独立 PR
（现状不影响读 / 修改；3-way split 收益有限，独立 PR 更稳）。

`db.init_db()` 内 `from .migrations import apply_all` 相对引用透明 work
（db.py 跟 migrations/ 都在 infrastructure/ 同一层）。

回归：PR-1 安全网 152 + supervisor 全套 + studio_server 44 + ... = **267/267
全绿**。route snapshot 不变（160 routes）。
